requirements.txt
agithub==2.2.2
asn1crypto==0.24.0
cachetools==3.1.1
celery==4.3.0
certifi==2018.11.29
cffi==1.11.5
cfssl==0.0.3b243
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
dj-stripe==2.1.1
Django~=2.2.5
django-allow-cidr==0.3.1
django-bootstrap4==1.1.1
django-extensions==2.2.1
django-json-widget==0.2.0
django-jsonfield==1.0.1
django-phonenumber-field==3.0.1
phonenumbers==8.10.23
django-prometheus==1.0.15
django-registration-redux==2.6
django-tagulous==0.14.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
dockerignore-generate==1.0.7
freezegun==0.3.11
GitPython==3.0.8
google-api-core==1.14.0
google-auth==1.6.3
google-cloud-bigquery==1.15.0
google-cloud-core==1.0.2
google-cloud-datastore==1.8.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jwt==0.6.1
markdown==3.1.1
MarkupSafe==1.1.0
mixpanel==4.5.0
netaddr==0.7.19
netifaces==0.10.9
prometheus-client==0.7.1
protobuf==3.9.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
redis==3.3.8
requests==2.21.0
rsa==4.0
sentry-sdk==0.7.14
sh==1.12.14
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
whitenoise==4.1.2
